modules/p189sign/pyproject.toml
[tool.poetry]
name = "p189sign"
version = "0.0.10"
description = "189 web disk signature."
authors = ["ChenyangGao <wosiwujm@gmail.com>"]
license = "MIT"
readme = "readme.md"
homepage = "https://github.com/ChenyangGao/p189client/tree/main/modules/p189sign"
repository = "https://github.com/ChenyangGao/p189client/tree/main/modules/p189sign"
keywords = ["python-189", "signature"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.12"
pycryptodome = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "p189sign"
